pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "nltk>=3.9.1",
    "openai>=1.63.0",
    "pillow>=11.1.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "streamlit>=1.42.0",
    "trafilatura>=2.0.0",
    "twilio>=9.4.4",
]
